Introduces hook blacklist, use measurement_name
Instead of offering seven different config keys for reports, let's just provide
a way to blacklist ActiveSupport::Notifications hooks. Makes the configuration
easier to digest for the users and if we ever introduce new subscribers easier
to extend for us.

Additionally we now report only one measurement ("rails" by default) with the
ActiveSupport::Notifications hook name as tag instead of 3.

This also switches the default for the SQL reporter from off to on. There are
probably more sites with a small amount of traffic than big sites that have to
worry about the performance impact the SQL normalisation in pure Ruby has. SQL
reports are wildly useful.