Drop Exceptions
To simplify the code base and in honor to the many Free Software exception
trackers (errbit, sentry etc.), who do this job in a much nicer way, drop
the exceptions feature completely.

We are sorry if you have relied on this and you now have to migrate to
something else for exception tracking, but this was such an invasive code-path.
We rather want to concentrate our resources on supporting more ActiveSupport
Instrumentation hooks and data instead of this.

Fixes #73